tests/wpt/web-platform-tests/navigation-api/navigation-history-entry/entries-array-equality.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_not_equals(navigation.entries(), navigation.entries());
}, "navigation.entries() should not return an identical object on repeated invocations");
</script>
